requirements.txt
# -------------------------------------------------
# Core LangChain (Modular v0.2.x)
# -------------------------------------------------
langchain==0.2.1
langchain-core==0.2.2
langchain-community==0.2.1
langchain-text-splitters==0.2.0

# -------------------------------------------------
# HuggingFace Embeddings (CPU-friendly)
# -------------------------------------------------
langchain-huggingface==0.0.3
sentence-transformers>=2.6.0
torch>=2.0.0,<2.3.0

# -------------------------------------------------
# Vector Store
# -------------------------------------------------
chromadb==0.5.0
langchain-chroma==0.1.4

# -------------------------------------------------
# Document Loaders
# -------------------------------------------------
pypdf==4.2.0
unstructured>=0.14.5
python-docx>=1.1.0

# -------------------------------------------------
# Google Gemini (LLM + Judge)
# -------------------------------------------------
langchain-google-genai==1.0.6
google-generativeai==0.5.4

# -------------------------------------------------
# Evaluation (Optional: kept lightweight)
# -------------------------------------------------
datasets==2.19.2
pandas==2.2.2

# -------------------------------------------------
# FastAPI Backend + UI
# -------------------------------------------------
fastapi>=0.110.0
uvicorn>=0.29.0
pydantic>=2.6.0
python-multipart>=0.0.9
jinja2>=3.1.3

# -------------------------------------------------
# Utilities
# -------------------------------------------------
python-dotenv==1.0.1
tqdm==4.66.4
